% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CLXPermHier}
\alias{CLXPermHier}
\title{Hierarchical Permutation to learn data-specific Gumbel distribution parameters}
\usage{
CLXPermHier(X, searchx, start, nt2 = 4L, minband = 20L, nperm = 100L)
}
\arguments{
\item{X:}{samples by covariates (genes)}

\item{searchx:}{candidate change point locations that will be tested.}

\item{start:}{the start location of the current segment}

\item{nperm:}{number of permutations}

\item{nt2:}{number of varied segment 2. Default is 4, which means we fix the length of segment 1 and vary the length of segment 2 for 4 times. The lengths for the varied segment 2s are equally spaced.}

\item{minband:}{minimum bandwidth of a segment. Default is 20.}
}
\description{
Hierarchical Permutation to learn data-specific Gumbel distribution parameters
}
