% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CLXstatPerm}
\alias{CLXstatPerm}
\title{Permutation to learn data-specific Gumbel distribution parameters}
\usage{
CLXstatPerm(X, searchx, start, nperm)
}
\arguments{
\item{X:}{samples by covariates (genes)}

\item{searchx:}{candidate change point locations that will be tested.}

\item{start:}{the start location of the current segment}

\item{nperm:}{number of permutations}
}
\description{
Permutation to learn data-specific Gumbel distribution parameters
}
