% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcpd_main_function.R
\name{covcpd}
\alias{covcpd}
\title{covcpd: detect covariance-based change points for scRNA-seq data}
\usage{
covcpd(
  X,
  nperm = 500,
  siglvl = 0.05,
  k = 30,
  maxseg = NULL,
  verbose = T,
  search_by = NULL,
  learnMode = c("linear", "GAM")[1]
)
}
\arguments{
\item{X}{raw count matrix, genes by cells}

\item{nperm}{number of permutations}

\item{siglvl}{significance level, default is 0.05}

\item{k}{the number of minimum cells in each segment, default is 30.}

\item{maxseg}{the maximum number of segments}

\item{verbose}{logical, whether to print out processing information. Default is TRUE.}

\item{search_by}{the number of cells between two consecutive candidate change points.}

\item{learnMode}{to learn the curve for Gumbel distribution parameters alpha and beta, we provide either "linear" or "GAM" method}
}
\value{
cps: the estimated change points location.
cps_order: the estimated change points location ordered by the order they were detected.
pvals: the p-values for the detected cps
candtau: candidate change point locations that were tested
pthresh: p-value threshold learned by permutation
minpdist: minimum p-value distribution learned by permuttaion.
}
\description{
covcpd: detect covariance-based change points for scRNA-seq data
}
