% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcpd_utils_function.R
\name{generateSigmaList}
\alias{generateSigmaList}
\title{Generate a list of covariance matrices.}
\usage{
generateSigmaList(
  nivec.list,
  u = c(-60:60)/100,
  ud = c(-100:-60, 60:100)/100,
  structure = "Identical S, Identical W",
  diffblk = NULL
)
}
\value{
a list of covariance matrices.
}
\description{
Generate a list of covariance matrices.
}
