% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{testCov_cpp}
\alias{testCov_cpp}
\title{calculate the Mn statistic from Cai et al. (2013)}
\usage{
testCov_cpp(X, Y, alpha = 0.05)
}
\arguments{
\item{X}{samples by covariates (genes)}

\item{Y}{samples by covariates (genes)}

\item{alpha}{significance level}
}
\description{
calculate the Mn statistic from Cai et al. (2013)
}
